Refactor API import and remove direct axios header manipulation
